#40: Fixed Sonar findings.
